Rename module: 'enums' -> 'enumerable'
